SCHEDULE BASED ON GREEDY  ALGORITHM
The first algorithm implemented for the solution of the job shop scheduling problem was by Greedy Algorithm, in this algorithm it was used the rule of Shortest Processing Time (SPT),  the idea of the rule is to make a sequence according to choosing the job that has the lowest processing time and to set it at first position and following under this assumption until to obtain the job with the higher processing time. This step was defined in the function greedy_sequence, and the following function was optimal_time. Once, it was determined the sequence of the jobs, it was essential to calculate the optimal time, starting time, idle time, wait time, and stop time

1.	Optimal Time: 55
2.	Optimal Sequence of: [7, 2, 3, 1, 4, 5, 6]