SCHEDULE BASED ON GENETIC  ALGORITHM
The second algorithm is focused on the implementation of a genetic algorithm, using the data from Table I again,  the fitness function was created according to the processing time, i.e., it was applied the greedy algorithm for calculating starting time, idle time, wait time, and stop time. The selection of the population was by tournament approach, the additional variables were population size of 400, 500 generations, mutation rate of 0.1, and tournament size of 5. The outcomes in the model were:
1.	Optimal Time: 53
2.	Optimal sequence of: [5, 7, 1, 3, 4, 2, 6]